+++ cmd/kubectl-testkube/commands/cloud/login.go
@@ -1,14 +1,10 @@
 package cloud
 
 import (
-	"context"
-
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/commands/common"
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/config"
-	"github.com/kubeshop/testkube/pkg/cloudlogin"
 	"github.com/kubeshop/testkube/pkg/ui"
 
-	"github.com/skratchdot/open-golang/open"
 	"github.com/spf13/cobra"
 )
 
@@ -21,23 +17,8 @@ func NewLoginCmd() *cobra.Command {
 		Short:   "Login to Testkube Cloud",
 		Run: func(cmd *cobra.Command, args []string) {
 			opts.CloudUris = common.NewCloudUris(opts.CloudRootDomain)
-			authUrl, tokenChan, err := cloudlogin.CloudLogin(context.Background(), opts.CloudUris.Auth)
-			ui.ExitOnError("cloud login", err)
-
 			ui.H1("Login")
-			ui.Paragraph("Your browser should open automatically. If not, please open this link in your browser:")
-			ui.Link(authUrl)
-			ui.Paragraph("(just login and get back to your terminal)")
-			ui.Paragraph("")
-
-			if ok := ui.Confirm("Continue"); !ok {
-				return
-			}
-
-			// open browser with login page and redirect to localhost
-			open.Run(authUrl)
-
-			token, refreshToken, err := uiGetToken(tokenChan)
+			token, refreshToken, err := common.LoginUser(opts.CloudUris.Auth)
 			ui.ExitOnError("getting token", err)
 
 			orgID := opts.CloudOrgId
