+++ cmd/kubectl-testkube/commands/common/client.go
@@ -76,9 +76,15 @@ func GetClient(cmd *cobra.Command) (client.Client, string, error) {
 
 		if cfg.CloudContext.ApiKey != "" && cfg.CloudContext.RefreshToken != "" {
 			var refreshToken string
-			token, refreshToken, err = cloudlogin.CheckAndRefreshToken(context.Background(), fmt.Sprintf("%s/idp", cfg.CloudContext.ApiUri), cfg.CloudContext.ApiKey, cfg.CloudContext.RefreshToken)
+			authURI := fmt.Sprintf("%s/idp", cfg.CloudContext.ApiUri)
+
+			token, refreshToken, err = cloudlogin.CheckAndRefreshToken(context.Background(), authURI, cfg.CloudContext.ApiKey, cfg.CloudContext.RefreshToken)
 			if err != nil {
-				return nil, "", fmt.Errorf("error checking token: %w", err)
+				// Error: failed refreshing, go thru login flow
+				token, refreshToken, err = LoginUser(authURI)
+				if err != nil {
+					return nil, "", fmt.Errorf("error logging in: %w", err)
+				}
 			}
 			if err := UpdateTokens(cfg, token, refreshToken); err != nil {
 				return nil, "", fmt.Errorf("error storing new token: %w", err)
