+++ cmd/kubectl-testkube/commands/common/helper.go
@@ -1,15 +1,19 @@
 package common
 
 import (
+	"context"
 	"fmt"
 	"os/exec"
 	"strings"
+	"time"
 
+	"github.com/skratchdot/open-golang/open"
 	"github.com/spf13/cobra"
 
 	"github.com/kubeshop/testkube/cmd/kubectl-testkube/config"
 	"github.com/kubeshop/testkube/internal/migrations"
 	cloudclient "github.com/kubeshop/testkube/pkg/cloud/client"
+	"github.com/kubeshop/testkube/pkg/cloudlogin"
 	"github.com/kubeshop/testkube/pkg/migrator"
 	"github.com/kubeshop/testkube/pkg/process"
 	"github.com/kubeshop/testkube/pkg/ui"
@@ -368,4 +372,46 @@ func PopulateCloudConfig(cfg config.Data, apiKey, orgId, envId, rootDomain strin
 	}
 
 	return cfg
+}
+
+func LoginUser(authUri string) (string, string, error) {
+	authUrl, tokenChan, err := cloudlogin.CloudLogin(context.Background(), authUri)
+	if err != nil {
+		return "", "", fmt.Errorf("cloud login: %w", err)
+	}
+	ui.H1("Login")
+	ui.Paragraph("Your browser should open automatically. If not, please open this link in your browser:")
+	ui.Link(authUrl)
+	ui.Paragraph("(just login and get back to your terminal)")
+	ui.Paragraph("")
+
+	if ok := ui.Confirm("Continue"); !ok {
+		return "", "", fmt.Errorf("login cancelled")
+	}
+
+	// open browser with login page and redirect to localhost
+	open.Run(authUrl)
+
+	idToken, refreshToken, err := uiGetToken(tokenChan)
+	if err != nil {
+		return "", "", fmt.Errorf("getting token")
+	}
+	return idToken, refreshToken, nil
+}
+
+func uiGetToken(tokenChan chan cloudlogin.Tokens) (string, string, error) {
+	// wait for token received to browser
+	s := ui.NewSpinner("waiting for auth token")
+
+	var token cloudlogin.Tokens
+	select {
+	case token = <-tokenChan:
+		s.Success()
+	case <-time.After(5 * time.Minute):
+		s.Fail("Timeout waiting for auth token")
+		return "", "", fmt.Errorf("timeout waiting for auth token")
+	}
+	ui.NL()
+
+	return token.IDToken, token.RefreshToken, nil
 }
